% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.pca.function.R
\name{filter.pca}
\alias{filter.pca}
\title{Filter out PCs colinear with other covariates.}
\usage{
\method{filter}{pca}(PCA, covs)
}
\arguments{
\item{PCA}{The output list from the prcomp() function.}

\item{covs}{A matrix/data.frame of covariates (rows = individuals, columns = covariates)}
}
\value{
A matrix of filtered PCs (not correlated with any covariates supplied by user)
}
\description{
Filter out PCs colinear with other covariates.
}
