% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GWAS.by.GLM.R
\name{GWAS.by.GLM}
\alias{GWAS.by.GLM}
\title{Perform GWA by GLM using PCs and/or Other Covariates}
\usage{
GWAS.by.GLM(gen.data, phen.data, gen.map, covs, corr.thresh, PCs)
}
\arguments{
\item{gen.data}{A numeric genotype matrix/data.frame (rows = taxa, columns = markers).}

\item{phen.data}{A phenotype matrix/data.frame (rows = taxa, column = trait scores). Only one trait is permitted.}

\item{gen.map}{A matrix/data.frame of SNP information (rows = SNPs, columns = c(SNP_ID, Chromosome, Position))}

\item{covs}{A matrix/data.frame of covariates (rows = taxa, columns = covariates)}

\item{corr.thresh}{A numerical value (0-1) indicating the maximum allowed correlation between any one covariate and any one PC}

\item{PCs}{A numerical value indicating the number of PCs to be included in the GLM equation as additional covariates to any supplied by the user}
}
\value{
A vector of p values. Each p value corresponding to one SNP tested.
}
\description{
Perform GWA by GLM using PCs and/or Other Covariates
}
