% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/<package-name>.R
\docType{package}
\name{GWAS.GLM}
\alias{GWAS.GLM}
\title{GWAS.GLM: A package for performing genome-wide associations by general linear modeling.}
\description{
The GWAS.by.GLM package consists of one main function that uses several side functions.
The main function is:
GWAS.by.GLM
}
