% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perform.GWA.GLM.R
\name{GWAS}
\alias{GWAS}
\title{Perform GWA by GLM}
\usage{
GWAS(gen.data, phen.data, covs = NULL, PCA, PCs = 3)
}
\arguments{
\item{gen.data}{A numeric genotype matrix/data.frame (rows = taxa, columns = markers).}

\item{phen.data}{A phenotype matrix/data.frame (rows = taxa, column = trait scores). Only one trait is permitted.}

\item{covs}{A matrix/data.frame of covariates (rows = taxa, columns = covariates)}

\item{PCA}{The output list from the prcomp() function}

\item{PCs}{A numerical value indicating the number of PCs to be included in the GLM equation as additional covariates to any supplied by the user}
}
\value{
A vector of p values. Each p value corresponding to one SNP tested.
}
\description{
Perform GWA by GLM
}
